+++ Common/Qt/Options/MultiHostTableOption.h
@@ -73,25 +73,25 @@ public:
 
 private:
     void replace_table();
-    void add_row(int row, const MultiHostTableOption::GameSlot& game, int& index_ref);
-    QComboBox* make_game_slot_box(QWidget& parent, int& row, int slot);
-    QComboBox* make_user_slot_box(QWidget& parent, int& row, int slot);
-    QLineEdit* make_skips_box(QWidget& parent, int& row, int skips);
-    QWidget* make_backup_save_box(QWidget& parent, int& row, bool backup_save);
-    QWidget* make_catchable_box(QWidget& parent, int& row, bool always_catchable);
-    QWidget* make_accept_FRs_box(QWidget& parent, int& row, bool accept_FRs);
-    QComboBox* make_move_slot_box(QWidget& parent, int& row, int move_slot);
-    QWidget* make_dynamax_box(QWidget& parent, int& row, bool dynamax);
-    QLineEdit* make_delay_box(QWidget& parent, int& row, const QString& post_raid_delay);
-    QPushButton* make_insert_button(QWidget& parent, int& row);
-    QPushButton* make_remove_button(QWidget& parent, int& row);
+    void add_row(size_t row, const MultiHostTableOption::GameSlot& game, size_t& index_ref);
+    QComboBox*      make_game_slot_box  (QWidget& parent, size_t& row, size_t slot);
+    QComboBox*      make_user_slot_box  (QWidget& parent, size_t& row, size_t slot);
+    QLineEdit*      make_skips_box      (QWidget& parent, size_t& row, size_t skips);
+    QWidget*        make_backup_save_box(QWidget& parent, size_t& row, bool backup_save);
+    QWidget*        make_catchable_box  (QWidget& parent, size_t& row, bool always_catchable);
+    QWidget*        make_accept_FRs_box (QWidget& parent, size_t& row, bool accept_FRs);
+    QComboBox*      make_move_slot_box  (QWidget& parent, size_t& row, size_t move_slot);
+    QWidget*        make_dynamax_box    (QWidget& parent, size_t& row, bool dynamax);
+    QLineEdit*      make_delay_box      (QWidget& parent, size_t& row, const QString& post_raid_delay);
+    QPushButton*    make_insert_button  (QWidget& parent, size_t& row);
+    QPushButton*    make_remove_button  (QWidget& parent, size_t& row);
 
     void update_table_height();
 
 private:
     MultiHostTableOption& m_value;
     QTableWidget* m_table;
-    std::vector<std::unique_ptr<int>> m_index_table;
+    std::vector<std::unique_ptr<size_t>> m_index_table;
 };
