+++ SerialPrograms/Source/PokemonSwSh/Inference/ShinyDetection/PokemonSwSh_ShinyFilters.h
@@ -18,7 +18,7 @@ namespace PokemonSwSh{
 struct BrightYellowLightFilter{
     size_t count = 0;
 
-    void operator()(FillMatrix::ObjectID& cell, const QImage& image, size_t x, size_t y){
+    void operator()(FillMatrix::ObjectID& cell, const QImage& image, int x, int y){
         QRgb pixel = image.pixel(x, y);
 //        int set = qRed(pixel) > 160 && qGreen(pixel) > 160 && qBlue(pixel) > 128;
         int set = (pixel & 0x00c0c080) == 0x00c0c080 ? 1 : 0;
@@ -30,7 +30,7 @@ struct BrightYellowLightFilter{
 struct BrightYellowLightFilterDebug{
     size_t count = 0;
 
-    void operator()(FillMatrix::ObjectID& cell, QImage& image, size_t x, size_t y){
+    void operator()(FillMatrix::ObjectID& cell, QImage& image, int x, int y){
         QRgb pixel = image.pixel(x, y);
 //        int set = qRed(pixel) > 160 && qGreen(pixel) > 160 && qBlue(pixel) > 128;
         int set = (pixel & 0x00c0c080) == 0x00c0c080 ? 1 : 0;
